Futher refactor ConstrainedGenerators. Update contexts and get rid of symbol
