+++ libs/constrained-generators/src/Constrained/Examples/Fold.hs
@@ -12,7 +12,6 @@ module Constrained.Examples.Fold where
 import Constrained.API
 import Constrained.Examples.List (Numbery)
 import Constrained.GenT (GE (..), catMessages, genFromGenT, inspect)
-import Constrained.Spec.ListFoldy (genListWithSize, predSpecPair)
 import Constrained.SumList
 import Data.String (fromString)
 import Prettyprinter (fillSep, punctuate, space)
@@ -130,7 +129,7 @@ pickProp = do
 -- | Build properties about calls to 'genListWithSize'
 testFoldSpec ::
   forall a.
-  (Foldy a, Random a, Integral a, TypeSpec a ~ NumSpec a) =>
+  (Foldy a, Random a, Integral a, TypeSpec a ~ NumSpec a, Arbitrary a, MaybeBounded a) =>
   Specification Integer ->
   Specification a ->
   Specification a ->
@@ -153,7 +152,13 @@ testFoldSpec size elemSpec total outcome = do
 -- | Generate a property from a call to 'pickAll'. We can test for success or failure using 'outcome'
 sumProp ::
   (Integral t, Random t, HasSpec t) =>
-  t -> t -> Specification t -> t -> Int -> Outcome -> Gen Property
+  t ->
+  t ->
+  Specification t ->
+  t ->
+  Int ->
+  Outcome ->
+  Gen Property
 sumProp smallest largest spec total count outcome = sumProp2 smallest largest (predSpecPair spec) total count outcome
 
 -- | Like SumProp, but instead of using a (Specification fn n) for the element predicate
@@ -161,7 +166,13 @@ sumProp smallest largest spec total count outcome = sumProp2 smallest largest (p
 --   using any Haskell function.
 sumProp2 ::
   (Show t, Integral t, Random t) =>
-  t -> t -> (String, t -> Bool) -> t -> Int -> Outcome -> Gen Property
+  t ->
+  t ->
+  (String, t -> Bool) ->
+  t ->
+  Int ->
+  Outcome ->
+  Gen Property
 sumProp2 smallest largest spec total count outcome = do
   (_, ans) <- pickAll smallest largest spec total count (Cost 0)
   let callString = parensList ["pickAll", show smallest, (fst spec), show total, show count]
