+++ libs/constrained-generators/src/Constrained/List.hs
@@ -1,7 +1,10 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE PolyKinds #-}
 {-# LANGUAGE QuantifiedConstraints #-}
@@ -11,13 +14,16 @@
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE UndecidableInstances #-}
 {-# LANGUAGE ViewPatterns #-}
 
 module Constrained.List where
 
+import Data.Foldable (fold)
 import Data.Functor.Const
 import Data.Kind
 import Data.Semigroup (Sum (..))
+import GHC.TypeLits
 
 -- | A heterogeneous list / an inductive tuple.
 -- We use this heavily to represent arguments to
@@ -32,6 +38,32 @@ infixr 5 :>
 
 deriving instance (forall a. Show (f a)) => Show (List f as)
 deriving instance (forall a. Eq (f a)) => Eq (List f as)
+
+type family Length as where
+  Length '[] = 0
+  Length (_ : as) = 1 + Length as
+
+type family as :! n where
+  '[] :! n = TypeError ('Text "Indexing into empty type-level list")
+  (a : as) :! 0 = a
+  (a : as) :! n = as :! (n - 1)
+
+class IndexOf n as where
+  at :: List f as -> f (as :! n)
+
+instance IndexOf 0 (a : as) where
+  at (a :> _) = a
+
+instance {-# OVERLAPPABLE #-} (IndexOf (n - 1) as, (as :! (n - 1)) ~ ((a : as) :! n)) => IndexOf n (a : as) where
+  at (_ :> as) = at @(n - 1) as
+
+toList :: (forall a. f a -> b) -> List f as -> [b]
+toList _ Nil = []
+toList f (x :> xs) = f x : toList f xs
+
+mapListC_ :: forall c f as b. All c as => (forall a. c a => f a -> b) -> List f as -> [b]
+mapListC_ _ Nil = []
+mapListC_ f (x :> xs) = f x : mapListC_ @c f xs
 
 mapList :: (forall a. f a -> g a) -> List f as -> List g as
 mapList _ Nil = Nil
@@ -55,13 +87,11 @@ mapMListC _ Nil = pure Nil
 mapMListC f (x :> xs) = (:>) <$> f x <*> mapMListC @c f xs
 
 foldMapList :: Monoid b => (forall a. f a -> b) -> List f as -> b
-foldMapList _ Nil = mempty
-foldMapList f (a :> as) = f a <> foldMapList f as
+foldMapList f = fold . toList f
 
 foldMapListC ::
   forall c as b f. (All c as, Monoid b) => (forall a. c a => f a -> b) -> List f as -> b
-foldMapListC _ Nil = mempty
-foldMapListC f (a :> as) = f a <> foldMapListC @c f as
+foldMapListC f = fold . mapListC_ @c f
 
 appendList :: List f as -> List f bs -> List f (Append as bs)
 appendList Nil bs = bs
