+++ libs/constrained-generators/src/Constrained/Properties.hs
@@ -18,29 +18,64 @@ module Constrained.Properties where
 import Constrained.API
 import Constrained.Base (
   AppRequires,
-  CList (..),
-  Context (..),
-  Ctx (Ctx, HOLE),
-  Mode (..),
+  BaseW (..),
+  HOLE (..),
   appTerm,
   (/>),
  )
-import Constrained.Conformance (monitorSpec)
-import Constrained.Core (Value (..), Var (..), unValue)
-import Constrained.GenT (GE (..), errorGE, fromGEDiscard, fromGEProp, strictGen)
+import Constrained.Conformance (
+  monitorSpec,
+ )
+import Constrained.Core (
+  Value (..),
+  Var (..),
+  unValue,
+ )
+import Constrained.GenT (
+  GE (..),
+  errorGE,
+  fromGEDiscard,
+  fromGEProp,
+  strictGen,
+ )
 import Constrained.List (
   All,
   FunTy,
   List (..),
+  ListCtx (..),
   TypeList,
+  fillListCtx,
   lengthList,
   listShape,
+  mapListCtxC,
+  mapMListC,
   uncurryList,
   uncurryList_,
  )
-import Constrained.Spec.ListFoldy (genInverse)
-import Constrained.Syntax (PolyCList (..))
-import Constrained.TheKnot
+import Constrained.NumSpec (
+  IntW (..),
+  NumOrdW (..),
+ )
+import Constrained.Spec.Map (
+  MapW (..),
+ )
+import Constrained.Spec.Set ()
+import Constrained.Spec.SumProd ()
+import Constrained.TheKnot (
+  BoolW (..),
+  ElemW (..),
+  EqW (..),
+  FunW (..),
+  ListW (..),
+  ProdW (..),
+  SizeW (..),
+  SumW (..),
+  genInverse,
+  mapSpec,
+  prettyPlan,
+  shrinkWithSpec,
+ )
+
 import qualified Data.List.NonEmpty as NE
 import Data.Map (Map)
 import Data.Typeable (Typeable, typeOf)
@@ -129,18 +164,18 @@ prop_conformEmpty a = QC.property $ conformsTo a (emptySpec @a)
 prop_univSound :: TestableFn -> QC.Property
 prop_univSound (TestableFn fn) =
   QC.label (show fn) $
-    QC.forAllShrinkBlind @QC.Property (genCtxForFn fn) QC.shrink $ \tc@(TestableCtx clist) ->
+    QC.forAllShrinkBlind QC.arbitrary QC.shrink $ \tc@(TestableCtx ctx) ->
       QC.forAllShrinkBlind QC.arbitrary QC.shrink $ \spec ->
         QC.counterexample ("\nfn ctx = " ++ showCtxWith fn tc) $
           QC.counterexample (show $ "\nspec =" <+> pretty spec) $
-            let sspec = simplifySpec (propagate (Context fn clist) spec)
+            let sspec = simplifySpec (propagate fn ctx spec)
              in QC.counterexample ("\n" ++ show ("propagate ctx spec =" /> pretty sspec)) $
                   QC.counterexample ("\n" ++ show (prettyPlan sspec)) $
                     QC.within 20_000_000 $
                       QC.forAllBlind (strictGen $ genFromSpecT sspec) $ \ge ->
                         fromGEDiscard $ do
                           a <- ge
-                          let res = runOnCtx clist fn a
+                          let res = uncurryList_ unValue (semantics fn) $ fillListCtx ctx $ \HOLE -> Value a
                           pure $
                             QC.counterexample ("\ngenerated value: a = " ++ show a) $
                               QC.counterexample ("\nfn ctx[a] = " ++ show res) $
@@ -170,81 +205,35 @@ specType MemberSpec {} = "MemberSpec"
 specType TypeSpec {} = "TypeSpec"
 specType TrueSpec {} = "TrueSpec"
 
--- ========================================================================
--- Combining function symbols and compatible contexts in interesting ways.
--- ========================================================================
-
--- | Builds a Term from A FunSym and an appropriate (List Term) of args
---   computed from the TestableCtx, using fillCList, which fills the HOLE with variable 'v0'
---  (uncurryList $ appTerm ProdW) (Lit True :> Lit False :> Nil)  ===> prod_ True False
-showCtxWith ::
-  forall s t as b.
-  (AppRequires s t as b, Fill as) =>
-  t s as b ->
-  TestableCtx as ->
-  String
-showCtxWith fn (TestableCtx ctx) = show $ pretty tm
-  where
-    tm :: Term b
-    tm =
-      uncurryList (appTerm fn) $
-        fillCList
-          (\case HOLE -> V $ Var 0 "v"; Ctx {} -> error ("Impossible, only HOLE has type (Ctx hole hole)"))
-          Lit
-          ctx
-
--- | Takes a CList, replaces the hole with 'x', and turns it into a (List Value as), then uses the
---   semantics of the function symbol 'fn' applying it to each item in the List in turn,
---   to get a resulting answer of type 'b'
-runOnCtx ::
-  forall s t as b x.
-  (AppRequires s t as b, Fill as) =>
-  CList 'Pre as x x ->
-  t s as b ->
-  x ->
-  b
-runOnCtx clist fn x = uncurryList_ unValue (semantics fn) (fillCList fillWithX Value clist)
-  where
-    fillWithX HOLE = Value x
-    fillWithX _ = error ("Impossible, only HOLE has type (Ctx hole hole)")
-
--- | Fill in the HOLE in a CList, and then turn it into a regular List
-class Fill as where
-  fillCList ::
-    forall f hole m.
-    (Ctx hole hole -> f hole) -> (forall a. Literal a => a -> f a) -> CList m as hole hole -> List f as
-
-instance Fill '[] where
-  fillCList _ _ End = Nil
-
-instance Fill as => Fill (a : as) where
-  fillCList f g (x :|> xs) = g x :> fillCList @as f g xs
-  fillCList f g (x :<| xs) = g x :> fillCList @as f g xs
-  fillCList f g (c :<> xs) = f c :> fillCList @as f g xs
-
 -- ============================================================
 -- An abstraction that hides everything about a function symbol
 -- But includes inside in the constraints, everything needed to
 -- use the function symbol
 
+showCtxWith ::
+  forall fn as b.
+  AppRequires fn as b =>
+  fn as b ->
+  TestableCtx as ->
+  String
+showCtxWith fn (TestableCtx ctx) = show tm
+  where
+    tm :: Term b
+    tm =
+      uncurryList (appTerm fn) $
+        fillListCtx (mapListCtxC @HasSpec (lit @_ . unValue) ctx) (\HOLE -> V $ Var 0 "v")
+
 data TestableFn where
   TestableFn ::
-    ( Fill as
-    , QC.Arbitrary (Specification b)
+    ( QC.Arbitrary (Specification b)
     , Typeable (FunTy as b)
-    , AppRequires s t as b
+    , AppRequires t as b
     ) =>
-    t s as b ->
+    t as b ->
     TestableFn
 
-genCtxForFn :: forall s t as b. AppRequires s t as b => t s as b -> QC.Gen (TestableCtx as)
-genCtxForFn _ = QC.arbitrary @(TestableCtx as)
-
-ex2 :: Prod (Value Bool) (Value Bool)
-ex2 = (uncurryList (semantics ProdW)) (Value True :> Value False :> Nil)
-
 instance Show TestableFn where
-  show (TestableFn (fn :: t s as b)) =
+  show (TestableFn (fn :: t as b)) =
     show fn ++ " :: " ++ show (typeOf (undefined :: FunTy as b))
 
 -- ===========================================================
@@ -253,87 +242,33 @@ instance Show TestableFn where
 
 data TestableCtx as where
   TestableCtx ::
-    ( HasSpec a
-    , All HasSpec as
-    ) =>
-    CList 'Pre as a a -> -- This is a Clist with only HOLE Ctx inside, the final 'a a' ensure this.
+    HasSpec a =>
+    ListCtx Value as (HOLE a) ->
     TestableCtx as
-
--- ===========================================================
--- QC.Arbitrary instances
--- ===========================================================
-
-genTestableCtx :: (All HasSpec as, TypeList as) => QC.Gen (TestableCtx as)
-genTestableCtx = QC.arbitrary
 
 instance forall as. (All HasSpec as, TypeList as) => QC.Arbitrary (TestableCtx as) where
   arbitrary = do
     let shape = listShape @as
     idx <- QC.choose (0, lengthList shape - 1)
-    genPre idx shape
-  shrink (TestableCtx clist) = map TestableCtx (shrinkCList clist)
+    go idx shape
+    where
+      go :: forall f as'. All HasSpec as' => Int -> List f as' -> QC.Gen (TestableCtx as')
+      go 0 (_ :> as) =
+        TestableCtx . (HOLE :?) <$> mapMListC @HasSpec (\_ -> Value <$> genFromSpec TrueSpec) as
+      go n (_ :> as) = do
+        TestableCtx ctx <- go (n - 1) as
+        TestableCtx . (:! ctx) . Value <$> genFromSpec TrueSpec
+      go _ _ = error "The impossible happened in Arbitrary for TestableCtx"
 
-genPre :: forall f as. All HasSpec as => Int -> List f as -> QC.Gen (TestableCtx as)
-genPre _ Nil = error ("The impossible happened, unless we chose 'idx' in the Arbitrary instance out of bounds")
-genPre n (_ :> xs) | n <= 0 = do Poly ys <- genPost xs; pure $ TestableCtx (HOLE :<> ys)
-genPre n ((_ :: f a) :> xs) =
-  do
-    TestableCtx ys <- genPre (n - 1) xs
-    x <- genFromSpec (TrueSpec @a)
-    pure $ (TestableCtx (x :|> ys))
+  shrink (TestableCtx ctx) = TestableCtx <$> shrinkCtx ctx
+    where
+      shrinkCtx :: forall c as'. All HasSpec as' => ListCtx Value as' c -> [ListCtx Value as' c]
+      shrinkCtx (c :? as) = (c :?) <$> go as
+      shrinkCtx (Value a :! ctx') = map ((:! ctx') . Value) (shrinkWithSpec TrueSpec a) ++ map (Value a :!) (shrinkCtx ctx')
 
-genPost :: forall f as. All HasSpec as => List f as -> QC.Gen (PolyCList as)
-genPost Nil = pure (Poly End)
-genPost ((_ :: f a) :> xs) =
-  do
-    Poly ys <- genPost xs
-    n <- genFromSpec (TrueSpec @a)
-    pure (Poly (n :<| ys))
-
-shrinkCList :: All HasSpec as => CList m as x y -> [CList m as x y]
-shrinkCList (HOLE :<> End) = []
-shrinkCList (x :|> HOLE :<> End) = [x' :|> HOLE :<> End | x' <- shrinkWithSpec TrueSpec x]
-shrinkCList (HOLE :<> x :<| End) = [HOLE :<> x' :<| End | x' <- shrinkWithSpec TrueSpec x]
-shrinkCList (a :|> b :|> HOLE :<> End) =
-  [ (a' :|> b' :|> HOLE :<> End)
-  | a' <- shrinkWithSpec TrueSpec a
-  , b' <- shrinkWithSpec TrueSpec b
-  ]
-shrinkCList (a :|> HOLE :<> c :<| End) =
-  [ (a' :|> HOLE :<> c' :<| End)
-  | a' <- shrinkWithSpec TrueSpec a
-  , c' <- shrinkWithSpec TrueSpec c
-  ]
-shrinkCList (HOLE :<> b :<| c :<| End) =
-  [ (HOLE :<> b' :<| c' :<| End)
-  | b' <- shrinkWithSpec TrueSpec b
-  , c' <- shrinkWithSpec TrueSpec c
-  ]
-shrinkCList (a :|> b :|> c :|> HOLE :<> End) =
-  [ (a' :|> b' :|> c' :|> HOLE :<> End)
-  | a' <- shrinkWithSpec TrueSpec a
-  , b' <- shrinkWithSpec TrueSpec b
-  , c' <- shrinkWithSpec TrueSpec c
-  ]
-shrinkCList (a :|> b :|> HOLE :<> d :<| End) =
-  [ (a' :|> b' :|> HOLE :<> d' :<| End)
-  | a' <- shrinkWithSpec TrueSpec a
-  , b' <- shrinkWithSpec TrueSpec b
-  , d' <- shrinkWithSpec TrueSpec d
-  ]
-shrinkCList (a :|> HOLE :<> c :<| d :<| End) =
-  [ (a' :|> HOLE :<> c' :<| d' :<| End)
-  | a' <- shrinkWithSpec TrueSpec a
-  , c' <- shrinkWithSpec TrueSpec c
-  , d' <- shrinkWithSpec TrueSpec d
-  ]
-shrinkCList (HOLE :<> b :<| c :<| d :<| End) =
-  [ (HOLE :<> b' :<| c' :<| d' :<| End)
-  | b' <- shrinkWithSpec TrueSpec b
-  , c' <- shrinkWithSpec TrueSpec c
-  , d' <- shrinkWithSpec TrueSpec d
-  ]
-shrinkCList clist = error ("CList length greater than 4\n" ++ show clist)
+      go :: forall as'. All HasSpec as' => List Value as' -> [List Value as']
+      go Nil = []
+      go (Value a :> as) = map ((:> as) . Value) (shrinkWithSpec TrueSpec a) ++ map (Value a :>) (go as)
 
 genTestableFn :: QC.Gen TestableFn
 genTestableFn = QC.arbitrary
@@ -385,9 +320,9 @@ instance QC.Arbitrary TestableFn where
 
 prop_mapSpec ::
   ( HasSpec a
-  , AppRequires s t '[a] b
+  , AppRequires t '[a] b
   ) =>
-  t s '[a] b ->
+  t '[a] b ->
   Specification a ->
   QC.Property
 prop_mapSpec funsym spec =
@@ -397,9 +332,9 @@ prop_mapSpec funsym spec =
 
 prop_propagateSpecSound ::
   ( HasSpec a
-  , AppRequires s t '[a] b
+  , AppRequires t '[a] b
   ) =>
-  t s '[a] b ->
+  t '[a] b ->
   b ->
   QC.Property
 prop_propagateSpecSound funsym b =
