+++ libs/constrained-generators/src/Constrained/Spec/Map.hs
@@ -27,25 +27,14 @@ module Constrained.Spec.Map where
 
 import Constrained.Base
 import Constrained.Conformance
-import Constrained.Core (Evidence (..), NonEmpty ((:|)), unionWithMaybe)
+import Constrained.Core
 import Constrained.GenT
 import Constrained.Generic (Prod (..))
 import Constrained.List
 import Constrained.NumSpec (cardinality, geqSpec, leqSpec, nubOrd)
-import Constrained.Spec.ListFoldy (
-  FoldSpec (..),
-  Foldy (..),
-  ListSpec (..),
-  adds,
-  combineFoldSpec,
-  conformsToFoldSpec,
-  elem_,
-  toPredsFoldSpec,
- )
 import Constrained.Spec.Set
-import Constrained.Spec.Size (Sized (..), maxSpec, sizeOf_)
 import Constrained.Spec.SumProd
-import Constrained.Syntax (forAll, genHint, letBind, unsafeExists)
+import Constrained.Syntax
 import Constrained.TheKnot
 import Control.Monad
 import Data.Foldable
@@ -57,7 +46,6 @@ import qualified Data.Map as Map
 import Data.Set (Set)
 import qualified Data.Set as Set
 import GHC.Generics
-import GHC.TypeLits
 import Prettyprinter
 import Test.QuickCheck hiding (Fun, Witness, forAll)
 
@@ -228,7 +216,7 @@ instance
                     (kvs' <> typeSpec (Cartesian (notEqualSpec k) mempty))
                     (Map.insert k v m)
                 _ ->
-                  genError
+                  genErrorNE
                     ( NE.fromList
                         [ "Failed to generate enough elements for the map:"
                         , "  m = " ++ show m
@@ -237,12 +225,12 @@ instance
                         , show $ "  simplifySpec kvs' = " <> pretty (simplifySpec kvs')
                         ]
                     )
-        explain1 ("  n = " ++ show n) $ go n kvs mempty
+        explain ("  n = " ++ show n) $ go n kvs mempty
   genFromTypeSpec (MapSpec mHint mustKeys mustVals size (simplifySpec -> kvs) foldSpec) = do
-    mustMap <- explain1 "Make the mustMap" $ forM (Set.toList mustKeys) $ \k -> do
+    !mustMap <- explain "Make the mustMap" $ forM (Set.toList mustKeys) $ \k -> do
       let vSpec = constrained $ \v -> satisfies (pair_ (Lit k) v) kvs
-      v <- explain1 (show $ "vSpec =" <+> pretty vSpec) $ genFromSpecT vSpec
-      pure (k, v)
+      v <- explain (show $ "vSpec =" <+> pretty vSpec) $ genFromSpecT vSpec
+      pure $ (k, v)
     let haveVals = map snd mustMap
         mustVals' = filter (`notElem` haveVals) mustVals
         size' = simplifySpec $ constrained $ \sz ->
@@ -254,12 +242,12 @@ instance
                 <> maxSpec (cardinality (fstSpec kvs)) -- (mapSpec FstW $ mapSpec ToGenericW kvs))
                 <> maxSpec (cardinalTrueSpec @k)
             )
-        foldSpec' = case foldSpec of
+        !foldSpec' = case foldSpec of
           NoFold -> NoFold
-          FoldSpec fn@(Fun symbol) sumSpec -> FoldSpec fn $ propagate (Context theAddFn (HOLE :<> mustSum :<| End)) sumSpec
+          FoldSpec fn@(Fun symbol) sumSpec -> FoldSpec fn $ propagate theAddFn (HOLE :? Value mustSum :> Nil) sumSpec
             where
               mustSum = adds (map (semantics symbol) haveVals)
-    let valsSpec =
+    let !valsSpec =
           ListSpec
             Nothing
             mustVals'
@@ -267,8 +255,8 @@ instance
             (simplifySpec $ constrained $ \v -> unsafeExists $ \k -> pair_ k v `satisfies` kvs)
             foldSpec'
 
-    restVals <-
-      explain
+    !restVals <-
+      explainNE
         ( NE.fromList
             [ "Make the restVals"
             , show $ "  valsSpec =" <+> pretty valsSpec
@@ -282,7 +270,7 @@ instance
         go m (v : restVals') = do
           let keySpec = notMemberSpec (Map.keysSet m) <> constrained (\k -> pair_ k (Lit v) `satisfies` kvs)
           k <-
-            explain
+            explainNE
               ( NE.fromList
                   [ "Make a key"
                   , show $ indent 4 $ "keySpec =" <+> pretty keySpec
@@ -319,14 +307,15 @@ instance
 -- Logic instances for
 ------------------------------------------------------------------------
 
-data MapW (sym :: Symbol) (dom :: [Type]) (rng :: Type) where
-  DomW :: forall k v. Ord k => MapW "dom_" '[Map k v] (Set k)
-  RngW :: forall k v. Ord k => MapW "rng_" '[Map k v] [v]
-  LookupW :: forall k v. Ord k => MapW "lookup_" '[k, Map k v] (Maybe v)
+data MapW (dom :: [Type]) (rng :: Type) where
+  DomW :: (HasSpec k, HasSpec v, IsNormalType k, IsNormalType v, Ord k) => MapW '[Map k v] (Set k)
+  RngW :: (HasSpec k, HasSpec v, IsNormalType k, IsNormalType v, Ord k) => MapW '[Map k v] [v]
+  LookupW ::
+    (HasSpec k, HasSpec v, IsNormalType k, IsNormalType v, Ord k) => MapW '[k, Map k v] (Maybe v)
 
-deriving instance Eq (MapW s dom rng)
+deriving instance Eq (MapW dom rng)
 
-mapSem :: MapW s d r -> FunTy d r
+mapSem :: MapW d r -> FunTy d r
 mapSem DomW = Map.keysSet
 mapSem RngW = Map.elems
 mapSem LookupW = Map.lookup
@@ -336,24 +325,19 @@ instance Semantics MapW where
 
 instance Syntax MapW
 
-instance Show (MapW s d r) where
+instance Show (MapW d r) where
   show DomW = "dom_"
   show RngW = "rng_"
   show LookupW = "lookup_"
 
 -- ============ DomW
 
-instance
-  (HasSpec k, HasSpec v, Ord k, IsNormalType v, IsNormalType k) =>
-  Logic "dom_" MapW '[Map k v] (Set k)
-  where
-  propagate ctxt (ExplainSpec [] s) = propagate ctxt s
-  propagate ctxt (ExplainSpec es s) = ExplainSpec es $ propagate ctxt s
-  propagate _ TrueSpec = TrueSpec
-  propagate _ (ErrorSpec msgs) = ErrorSpec msgs
-  propagate (Context DomW (HOLE :<> End)) (SuspendedSpec v ps) =
-    constrained $ \v' -> Let (App DomW (v' :> Nil)) (v :-> ps)
-  propagate (Context DomW (HOLE :<> End)) spec =
+instance Logic MapW where
+  propagate f ctxt (ExplainSpec es s) = explainSpec es $ propagate f ctxt s
+  propagate _ _ TrueSpec = TrueSpec
+  propagate _ _ (ErrorSpec msgs) = ErrorSpec msgs
+  propagate f ctx (SuspendedSpec v ps) = constrained $ \v' -> Let (App f (fromListCtx ctx v')) (v :-> ps)
+  propagate DomW (Unary HOLE) spec =
     case spec of
       MemberSpec (s :| []) ->
         typeSpec $
@@ -368,80 +352,39 @@ instance
             (constrained $ \kv -> satisfies (fst_ kv) elemspec)
             NoFold
       _ -> ErrorSpec (NE.fromList ["Dom on bad map spec", show spec])
-  propagate ctx _ =
-    ErrorSpec $ pure ("Logic instance for DomW with wrong number of arguments. " ++ show ctx)
-
-  mapTypeSpec DomW (MapSpec _ mustSet _ sz kvSpec _)
-    | Evidence <- prerequisites @(Map k v) = typeSpec $ SetSpec mustSet (fstSpec @k @v kvSpec) sz
-
--- ============ RngW
-
-instance
-  (HasSpec k, HasSpec v, Ord k, IsNormalType v, IsNormalType k) =>
-  Logic "rng_" MapW '[Map k v] [v]
-  where
-  propagate ctxt (ExplainSpec [] s) = propagate ctxt s
-  propagate ctxt (ExplainSpec es s) = ExplainSpec es $ propagate ctxt s
-  propagate _ TrueSpec = TrueSpec
-  propagate _ (ErrorSpec msgs) = ErrorSpec msgs
-  propagate (Context RngW (HOLE :<> End)) (SuspendedSpec v ps) =
-    constrained $ \v' -> Let (App RngW (v' :> Nil)) (v :-> ps)
-  propagate (Context RngW (HOLE :<> End)) spec
-    | Evidence <- prerequisites @(Map k v) =
-        case spec of
-          TypeSpec (ListSpec listHint must size elemspec foldspec) [] ->
-            typeSpec $
-              MapSpec
-                listHint
-                Set.empty
-                must
-                size
-                (constrained $ \kv -> satisfies (snd_ kv) elemspec)
-                foldspec
-          -- NOTE: you'd think `MemberSpec [r]` was a safe and easy case. However, that
-          -- requires not only that the elements of the map are fixed to what is in `r`,
-          -- but they appear in the order that they are in `r`. That's
-          -- very difficult to achieve!
-          _ -> ErrorSpec (NE.fromList ["Rng on bad map spec", show spec])
-  propagate ctx _ =
-    ErrorSpec $ pure ("Logic instance for RngW with wrong number of arguments. " ++ show ctx)
-
-  mapTypeSpec RngW (MapSpec _ _ mustList sz kvSpec foldSpec)
-    | Evidence <- prerequisites @(Map k v) =
-        typeSpec $ (ListSpec Nothing mustList sz (sndSpec @k @v kvSpec) foldSpec)
-
--- ============ LookupW
-
-instance
-  (HasSpec k, HasSpec v, IsNormalType v, IsNormalType k) =>
-  Logic "lookup_" MapW '[k, Map k v] (Maybe v)
-  where
-  propagate ctxt (ExplainSpec [] s) = propagate ctxt s
-  propagate ctxt (ExplainSpec es s) = ExplainSpec es $ propagate ctxt s
-  propagate _ TrueSpec = TrueSpec
-  propagate _ (ErrorSpec msgs) = ErrorSpec msgs
-  propagate (Context LookupW (HOLE :<> x :<| End)) (SuspendedSpec v ps) =
-    constrained $ \v' -> Let (App LookupW (v' :> Lit x :> Nil)) (v :-> ps)
-  propagate (Context LookupW (x :|> HOLE :<> End)) (SuspendedSpec v ps) =
-    constrained $ \v' -> Let (App LookupW (Lit x :> v' :> Nil)) (v :-> ps)
-  propagate (Context LookupW (k :|> HOLE :<> End)) spec
-    | Evidence <- prerequisites @(Map k v) =
-        constrained $ \m ->
-          [Assert $ Lit k `member_` dom_ m | not $ Nothing `conformsToSpec` spec]
-            ++ [ forAll m $ \kv ->
-                  letBind (fst_ kv) $ \k' ->
-                    letBind (snd_ kv) $ \v ->
-                      whenTrue (Lit k ==. k') $
-                        -- TODO: What you want to write is `cJust_ v `satisfies` spec` but we can't
-                        -- do that because we don't have access to `IsNormalType v` here. When
-                        -- we refactor the `IsNormalType` machinery we will be able to make
-                        -- this nicer.
-                        case spec of
-                          MemberSpec as -> Assert $ v `elem_` Lit [a | Just a <- NE.toList as]
-                          TypeSpec (SumSpec _ _ vspec) cant ->
-                            v `satisfies` (vspec <> notMemberSpec [a | Just a <- cant])
-               ]
-  propagate (Context LookupW (HOLE :<> m :<| End)) spec =
+  propagate RngW (Unary HOLE) spec =
+    case spec of
+      TypeSpec (ListSpec listHint must size elemspec foldspec) [] ->
+        typeSpec $
+          MapSpec
+            listHint
+            Set.empty
+            must
+            size
+            (constrained $ \kv -> satisfies (snd_ kv) elemspec)
+            foldspec
+      -- NOTE: you'd think `MemberSpec [r]` was a safe and easy case. However, that
+      -- requires not only that the elements of the map are fixed to what is in `r`,
+      -- but they appear in the order that they are in `r`. That's
+      -- very difficult to achieve!
+      _ -> ErrorSpec (NE.fromList ["Rng on bad map spec", show spec])
+  propagate LookupW (Value k :! Unary HOLE) spec =
+    constrained $ \m ->
+      [Assert $ Lit k `member_` dom_ m | not $ Nothing `conformsToSpec` spec]
+        ++ [ forAll m $ \kv ->
+              letBind (fst_ kv) $ \k' ->
+                letBind (snd_ kv) $ \v ->
+                  whenTrue (Lit k ==. k') $
+                    -- TODO: What you want to write is `cJust_ v `satisfies` spec` but we can't
+                    -- do that because we don't have access to `IsNormalType v` here. When
+                    -- we refactor the `IsNormalType` machinery we will be able to make
+                    -- this nicer.
+                    case spec of
+                      MemberSpec as -> Assert $ v `elem_` Lit [a | Just a <- NE.toList as]
+                      TypeSpec (SumSpec _ _ vspec) cant ->
+                        v `satisfies` (vspec <> notMemberSpec [a | Just a <- cant])
+           ]
+  propagate LookupW (HOLE :? Value m :> Nil) spec =
     if Nothing `conformsToSpec` spec
       then notMemberSpec [k | (k, v) <- Map.toList m, not $ Just v `conformsToSpec` spec]
       else
@@ -453,8 +396,9 @@ instance
               , "  " ++ show spec
               ]
           )
-  propagate ctx _ =
-    ErrorSpec $ pure ("Logic instance for LookupW with wrong number of arguments. " ++ show ctx)
+
+  mapTypeSpec DomW (MapSpec _ mustSet _ sz kvSpec _) = typeSpec $ SetSpec mustSet (fstSpec kvSpec) sz
+  mapTypeSpec RngW (MapSpec _ _ mustList sz kvSpec foldSpec) = typeSpec $ (ListSpec Nothing mustList sz (sndSpec kvSpec) foldSpec)
 
 ------------------------------------------------------------------------
 -- Syntax
